# grpc-bff/bff-api/src/swagger.yaml
openapi: 3.0.1
info:
  title: Users Service REST API
  version: v1
servers:
  - description: 'Local'
    url: 'http://localhost'
tags:
  - description: Users
    name: users
paths:
  /bff/users:
    get:
      description: This request retrieves all users matching the search string.
      operationId: getUsers
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
          description: OK
        '401':
          description: Unauthorized
        '403':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorInfo'
          description: Forbidden
      tags:
        - users
components:
  schemas:
    Page:
      type: object
      properties:
        numItems:
          type: integer
          format: int32
          example: 50
        objects:
          type: array
          items:
            $ref: '#/components/schemas/User'
        totalNumItems:
          type: integer
          format: int64
          example: 1234
    User:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
          example: John
        lastname:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@mail.org
        roles:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: object
            properties:
              tech:
                type: string
              seniority:
                type: string
        education:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              level:
                type: string
        createdAt:
          type: string
          format: date
          example: 2017-12-06
        updatedAt:
          type: string
          format: date
          example: 2017-12-06
    ErrorInfo:
      type: object
      properties:
        exception:
          type: string
        message:
          type: string